[LayoutNG] SVGs with percentage dimensions on child-initiated relayouts
On child-initiated re-layouts, SVG elements with percentage dimensions
overflow auto-sized parents when LayoutNG is enabled. The reason being
that ObjectIsRelayoutBoundary() treats all SVGs as layout roots upon
re-layout, even when the width and height are not fixed.

With LayoutNG enabled, this becomes a problem because
BoxLayoutExtraInput does not get set when LayoutSVGRoot is a layout
root, leading to a miscalculation in sizing.

One option would be to set BoxLayoutExtraInput for LayoutSVGRoots
when they are layout roots. However, in order to keep SVG behavior
consistent with other objects, this change fixes the above issue
by preventing layout root optimization altogether for SVG elements
that do not have a fixed height and width.

Bug: 1016759
Change-Id: Id0ccdafb20e5ce3a1b58be3577ca74dccc0c20c5