Rename mojom-side transforms on input sources for consistency
There should be no functional changes introduced by this CL.

Renames XRInputSource-related data in mojom structs. This CL does not
attempt to fix names of other variables / functions that are used to
compute the data that will then be set on the mojom structs - this will
have to happen in a follow-up (mostly on device side, blink usage should
already be consistent).

Change-Id: I1b4888158a7ed57cecce7de87634e639cf23d6ad